--- Dockerfile
@@ -1,12 +1,0 @@
-FROM golang:1.24.5-alpine as BUILDER
-WORKDIR /app
-COPY go.mod ./
-COPY go.sum ./
-RUN go mod download
-COPY . .
-RUN go build -o app ./cmd
-
-FROM alpine
-WORKDIR /app
-COPY --from=builder /app/app .
-CMD ["./app"]
