+++ apps/gateway/internal/userNoteRepo/userNoteRepo.go
@@ -1,0 +1,14 @@
+package usernoterepo
+
+import (
+	"time"
+)
+
+type UserNote struct {
+	ID        int
+	AccountID int
+	Title     string
+	Body      string
+	CreatedAt time.Time
+	UpdatedAt time.Time
+}
