+++ apps/notes-svc/internal/httpServer/testdata/fixtures/note.sql
