--- configs/config.yaml
@@ -1,8 +1,0 @@
-server:
-  port: "5680"
-db:
-  url: "postgres://postgres:bear@db:5432/notedb"
-logger:
-  timestamp: "2006-01-02 15:04:05"
-  formatlevel: "%-6s"
-  level: "debug"
